projeto Pooparte2

+++ src/main/java/br/com/exercicio_26set/Desenvolvedor.java
@@ -5,7 +5,7 @@ public class Desenvolvedor extends Funcionario{
     private int horasExtras;
     private Double valorHoraExtra;
 
-    public Desenvolvedor(String nome,String matricula,Double salarioBase, int horasExtras, Double valorHoraExtra){
+    public Desenvolvedor(String nome, String matricula, int salarioBase, int horasExtras, Double valorHoraExtra){
         super(nome,matricula,salarioBase);
         this.horasExtras = horasExtras;
         this.valorHoraExtra = valorHoraExtra;
